worker/requirements.txt
pymongo>=4.8.0
python-dotenv>=1.0.0
redis>=5.0.0
